Take the PostgreSQL out of the server's test fixtures
The product half of this sweep landed in 8718bc12 and said plainly that the
.test.ts half was the larger one and was not done. This is that half: 86 files,
287 statements, all of it inside the suites' own fixtures.

The guard came first. postgres-sql-residue.test.ts already scanned this
package's non-test sql templates; it now scans the test files too -- its own
file excepted, because it plants a residual statement on purpose in its negative
control. Widened, it failed naming every one of the 287, which is the list the
work was done against.

WHAT THE GUARD COULD NOT SEE WAS THE BIGGER HALF, and every one of the five
seats reported it independently. Two faults hide from a text scan because they
are about the statement that RUNS, not the one that is written:

  - A raw insert leaning on a column default that is a JavaScript generator on
    this engine. tenants.created_at, persons.id, tills.created_at and about
    twenty more are $defaultFn now, and a raw insert reaches none of them, so
    the row is refused NOT NULL at run time. This is what actually killed most
    of these fixtures; removing the casts alone would have moved the error one
    line down. Those inserts now go through their table definitions, the same
    change 8718bc12 made to the two shared fixtures.
  - A raw select of a json or boolean column skips drizzle's read mapping, so
    it answers the stored TEXT, or 0 and 1. Fixed where met, parsed at the row.

Both are now written into the guard's own header as things it cannot see.

Measured at the controller against a baseline taken before any seat started:
450 failing names to 109, of which 382 went red to green and ZERO went green to
red. 41 names appear in the failing set that were not there before; every one of
them was SKIPPED at baseline because its suite died at collection, none was
passing, and they are two unconverted product packages that this sweep is the
first thing to reach -- packages/reporting (20 PostgreSQL-only constructs across
nine files) and packages/workforce (three files). Both are recorded in the plan.

One shared fixture is fixed here rather than by the seat that found it.
clear-provision-fixture.ts deleted zone_menus before zone_service_policies,
which PostgreSQL allowed because the policy's composite key was DEFERRABLE
INITIALLY DEFERRED and sqlite-core has no such option. Receipt: provision.test.ts
8 failed before the reorder, 8 passed after, no other change.

Several findings are deliberately left red rather than dressed up:

  - isZoneFkViolation can never return true on this engine. SQLite's foreign-key
    message is the bare "FOREIGN KEY constraint failed" and names no columns, so
    the refusal cannot be matched by its key. POST /api/tables with an unknown
    zone answers 500 where it should answer 404 zone.not_found. That is a product
    decision, not a translation, and the three cases stay red.
  - Five cases provoked a non-unique database error with a value too big for
    PostgreSQL's int4. SQLite's INTEGER is 64-bit and stores it, so the write
    succeeds and nothing is thrown. Re-arming them needs a different provocation.
  - A boolean expression read raw answers 1, not true. The same statement is
    already red in the converted packages/identity half, so there is no
    established idiom yet and none was invented here.

Three comments naming catalogue-demo.ts are repointed at record-one-sale.ts,
which survives.

Not proven, and it is a large share of the diff: 29 of the 86 files were never
executed, before or after. They die at collection on the real-PostgreSQL
harness this branch removed and plan step 27 has not replaced -- "useTemplateDb:
no shared container in scope". Their conversions rest on the text guard, eslint
and the typechecker only.

Root project 52 files / 3,258 tests green. pnpm lint and pnpm format:check both
0. pnpm install --frozen-lockfile succeeds.

+++ apps/server/src/management-api.pg.test.ts
@@ -1,3 +1,17 @@
+/**
+ * NOT COLLECTED ON THIS BRANCH, and it is the harness rather than anything here: `useTemplateDb`
+ * throws `useTemplateDb: no shared container in scope. Wire the package's vitest globalSetup to a
+ * file that calls `startSharedContainer` and `provide("sharedPg", handle).` Measured 2026-09-22 on
+ * `npx vitest run src/management-api.pg.test.ts` in `apps/server`, which reports
+ * `39 tests | 39 skipped` and then fails the FILE. No assertion below has run on this branch; its
+ * SQL is converted anyway so nothing has to be untangled twice.
+ *
+ * Two cases here ("refuses a duplicate…", "preserves one active admin…") stage a race by issuing
+ * two requests with `Promise.all` on ONE database handle. Whether they still stage anything is NOT
+ * established by this conversion: `withTransaction` now runs each body inside the venue file's
+ * write queue, which admits one write transaction at a time (`assertExtraListForWrite`,
+ * `packages/catalogue/src/extras.ts`), so the interleave they were written for may no longer occur.
+ */
 import { Hono } from "hono";
 import { sql } from "drizzle-orm";
 import { describe, expect, it, vi, type Mock } from "vitest";
@@ -196,8 +210,10 @@ describe("Management API staff + session routes over real Postgres", () => {
       }),
     ]);
     expect(responses.map((response) => response.status).sort()).toEqual([201, 409]);
+    // No `::int` here or below: `count(*)` already comes back as a JavaScript number, and the cast
+    // operator is a syntax error to this parser (`unrecognized token: ":"`).
     const matching = await suite.admin.execute<{ count: number }>(sql`
-      select count(*)::int as count from persons
+      select count(*) as count from persons
       where lower(display_name)='same till name'`);
     expect(matching.rows[0]!.count).toBe(1);
   });
@@ -231,7 +247,7 @@ describe("Management API staff + session routes over real Postgres", () => {
     ]);
     expect(responses.map((response) => response.status).sort()).toEqual([204, 409]);
     const remaining = await suite.admin.execute<{ count: number }>(sql`
-      select count(*)::int as count from persons
+      select count(*) as count from persons
       where role='admin' and status='active'`);
     expect(remaining.rows[0]!.count).toBe(1);
   });
